Moved the template matching to a specific file to separate it from the rest of the groundwater detection
